- Fixed typecasting in save SQL queries
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@1341 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/PlayerSkills.cpp
@@ -81,7 +81,7 @@ void PlayerSkills::save() {
 		else {
 			query << ",(";
 		}
-		query << player->getId() << "," << iter->first << "," << iter->second.level << "," << iter->second.maxlevel << ")";
+		query << player->getId() << "," << iter->first << "," << (int16_t) iter->second.level << "," << (int16_t) iter->second.maxlevel << ")";
 	}
 	query.exec();
 }
